Filter regex lookups by prescanning the regex input string and quickly ruling out inputs that can't match.
This results in a ~3x speedup on the 10k cluster test from
https://github.com/envoyproxy/envoy/issues/2063 .  Though it still
takes too long to startup IMO.

This shouldn't really help at all; I didn't convert to a map because I
was unsure of what to enter into the map.  I could do an
upper-bound/lower-bound type search though.  But the optimization in
this commit is more of a peephole tweek and I'm surprised it helped at
all.  I wonder if this speeks to how suboptimal std::regex is for our
application.

I have two higher level thoughts:
  1. does the std::regex library vasty underperform RE2?  I looked for an extant
     benchmark but did not find one.  I might try porting to RE2 just to see if
     that improves things much more.
  2. are we calling regex way too much in the startup flow?

Signed-off-by: Joshua Marantz <jmarantz@google.com>