Convert a few more regexes to tokenizers.
Removed the map, and any need to find a descriptor by name (which I
think was for tests only).  This allowed me to cover the same variable
name with multiple alternative rules.  We could actually skip looking for
a variable once it's been filled in, but I didn't do that yet.

This brings the startup timing to about 4.5 seconds.  Next step is to
form a new map of tokenizers, mapping from the first token to a list of
patterns that might match it, rather than linearly scanning all of them
all the time.

Signed-off-by: Joshua Marantz <jmarantz@google.com>