Don't create the match/removal bits until we have a confirmed success.
This improves speed very slightly, but more importantly the previous behavior had
a bug becaues it would add to the remove_characters IntervalSet on an early match,
even if it was going to fail later on.  We should get past the failure-exits before
doing any mutations visible outside the method.

Signed-off-by: Joshua Marantz <jmarantz@google.com>