Use a vector<pair<metric,value>> for temporarily collecting and sorting stats during admin queries.
Previously it was a map<string,value> with the stat name duplicated in the string.

This saves 2.4% of memory highwater after requesting /stats on 10001
clusters, reducing 810786816 bytes to 790863872.

The reason it doesn't save more is that still the admin implementation
buffers up the entire string responsse in memory, and so due to ahmdal's
law the percent benefiit of this one change is not as much as hoped.

I think combining this with a change to stream admin output when
possible would make this change look more impactful.

Signed-off-by: Joshua Marantz <jmarantz@google.com>